+++ src/server/game/Pools/PoolMgr.cpp
@@ -300,7 +300,7 @@ void PoolGroup<T>::SpawnObject(SpawnedPoolData& spawns, uint32 limit, uint32 tri
         // roll objects to be spawned
         if (!ExplicitlyChanced.empty())
         {
-            float roll = (float)rand_chance();
+            float roll = rand_chance();
 
             for (PoolObject& obj : ExplicitlyChanced)
             {
